+++ src/Frankenstein/animation.cpp
@@ -8,10 +8,10 @@
 * param c - startingRow, the row this animation starts at on the spritesheet
 * param d - name of this animation */
 Animation::Animation(GLuint image, int a, int b, int c, const std::string& d, const int facingDir): m_image(image),
-					m_numberOfFrames(a), m_numberOfRows(b), m_frameDivision(1.0/m_numberOfFrames),
-					m_rowDivision(1.0/m_numberOfRows), m_startingRow(c), m_currentFrame(0),
-					m_currentRow(m_startingRow), m_s1(0.0), m_s2(m_frameDivision), m_t1(m_currentRow*m_rowDivision), 
-					m_t2((m_currentRow*m_rowDivision) + m_rowDivision), m_name(d), m_facingDirection(facingDir){}
+                   m_numberOfFrames(a), m_numberOfRows(b), m_frameDivision(1.0f/static_cast<float>(m_numberOfFrames)),
+				   m_rowDivision(1.0f/static_cast<float>(m_numberOfRows)), m_startingRow(c), m_currentFrame(0),
+				   m_currentRow(m_startingRow), m_s1(0.0f), m_s2(m_frameDivision), m_t1(m_currentRow*m_rowDivision), 
+				   m_t2((m_currentRow*m_rowDivision) + m_rowDivision), m_name(d), m_facingDirection(facingDir){}
 
 Animation::Animation(AnimationParameters& params, SpriteSheetParameters& spriteSheetParams, const std::string& name, const int facingDirection) :
 
@@ -21,8 +21,8 @@ Animation::Animation(AnimationParameters& params, SpriteSheetParameters& spriteS
 	m_numberOfRows(params.rowsInAnimation),
 	m_startingRow(params.startingRow),
 	m_startingCol(params.startingCol),
-	m_frameDivision(1.0 / (m_spriteSheetInfo.m_spriteSheetWidth / m_spriteSheetInfo.m_frameWidth)),
-	m_rowDivision(1.0 / (m_spriteSheetInfo.m_spriteSheetHeight / m_spriteSheetInfo.m_frameHeight)),
+   m_frameDivision(1.0f / static_cast<float>(m_spriteSheetInfo.m_spriteSheetWidth / m_spriteSheetInfo.m_frameWidth)),
+   m_rowDivision(1.0f / static_cast<float>(m_spriteSheetInfo.m_spriteSheetHeight / m_spriteSheetInfo.m_frameHeight)),
 	m_currentFrame(m_startingCol),
 	m_currentRow(m_startingRow),
 	m_s1(0.0),
@@ -41,8 +41,8 @@ Animation::Animation(GLuint image, int a, int b, int c, int c2, SpriteSheetParam
 	m_startingRow(c),
 	m_startingCol(c2),
 	m_spriteSheetInfo(spriteSheetParams),
-	m_frameDivision(1.0 / (m_spriteSheetInfo.m_spriteSheetWidth / m_spriteSheetInfo.m_frameWidth)),
-	m_rowDivision(1.0 / (m_spriteSheetInfo.m_spriteSheetHeight / m_spriteSheetInfo.m_frameHeight)),
+   m_frameDivision(1.0f / static_cast<float>(m_spriteSheetInfo.m_spriteSheetWidth / m_spriteSheetInfo.m_frameWidth)),
+   m_rowDivision(1.0f / static_cast<float>(m_spriteSheetInfo.m_spriteSheetHeight / m_spriteSheetInfo.m_frameHeight)),
 	m_currentFrame(m_startingCol),
 	m_currentRow(m_startingRow),
 	m_s1(0.0),
@@ -55,7 +55,7 @@ Animation::Animation(GLuint image, int a, int b, int c, int c2, SpriteSheetParam
 
 void Animation::nextFrame(float frameTime, float animFPS){
 	// Calculate how many frames to jump ahead
-	m_currentFrame += frameTime * animFPS;
+   m_currentFrame += static_cast<int>(frameTime * animFPS);
 
 	// Keep currentFrame in range
 	if (m_currentFrame >= m_numberOfFrames)
